backend/requirements.txt
flask
flask-cors
google-generativeai
firebase-admin
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
pytest
pytest-mock
